Add some types - 93 problems left!
