+++ .vscode/settings.json
@@ -1,6 +1,10 @@
 {
     "cSpell.words": [
         "armcord",
-        "smartfrigde"
+        "armcordinternal",
+        "Autogain",
+        "modloader",
+        "smartfrigde",
+        "vaapi"
     ]
 }
