+++ src/common/lang.ts
@@ -6,8 +6,8 @@ export function setLang(language: string): void {
     if (!fs.existsSync(langConfigFile)) {
         fs.writeFileSync(langConfigFile, "{}", "utf-8");
     }
-    let rawdata = fs.readFileSync(langConfigFile, "utf-8");
-    let parsed = JSON.parse(rawdata);
+    let rawData = fs.readFileSync(langConfigFile, "utf-8");
+    let parsed = JSON.parse(rawData);
     parsed.lang = language;
     let toSave = JSON.stringify(parsed, null, 4);
     fs.writeFileSync(langConfigFile, toSave, "utf-8");
@@ -19,8 +19,8 @@ export function getLang(object: string): Promise<string> {
             const userDataPath = app.getPath("userData");
             const storagePath = path.join(userDataPath, "/storage/");
             const langConfigFile = `${storagePath}lang.json`;
-            let rawdata = fs.readFileSync(langConfigFile, "utf-8");
-            let parsed = JSON.parse(rawdata);
+            let rawData = fs.readFileSync(langConfigFile, "utf-8");
+            let parsed = JSON.parse(rawData);
             language = parsed.lang;
         } catch (_e) {
             console.log("Language config file doesn't exist. Fallback to English.");
@@ -34,13 +34,13 @@ export function getLang(object: string): Promise<string> {
     if (!fs.existsSync(langPath)) {
         langPath = path.join(import.meta.dirname, "../", "/assets/lang/en-US.json");
     }
-    let rawdata = fs.readFileSync(langPath, "utf-8");
-    let parsed = JSON.parse(rawdata);
+    let rawData = fs.readFileSync(langPath, "utf-8");
+    let parsed = JSON.parse(rawData);
     if (parsed[object] == undefined) {
         console.log(`${object} is undefined in ${language}`);
         langPath = path.join(import.meta.dirname, "../", "/assets/lang/en-US.json");
-        rawdata = fs.readFileSync(langPath, "utf-8");
-        parsed = JSON.parse(rawdata);
+        rawData = fs.readFileSync(langPath, "utf-8");
+        parsed = JSON.parse(rawData);
         return parsed[object];
     } else {
         return parsed[object];
@@ -52,8 +52,8 @@ export function getLangName(): string {
             const userDataPath = app.getPath("userData");
             const storagePath = path.join(userDataPath, "/storage/");
             const langConfigFile = `${storagePath}lang.json`;
-            let rawdata = fs.readFileSync(langConfigFile, "utf-8");
-            let parsed = JSON.parse(rawdata);
+            let rawData = fs.readFileSync(langConfigFile, "utf-8");
+            let parsed = JSON.parse(rawData);
             language = parsed.lang;
         } catch (_e) {
             console.log("Language config file doesn't exist. Fallback to English.");
