+++ src/tray.ts
@@ -1,11 +1,11 @@
-import * as fs from "fs";
+import fs from "fs";
 import {Menu, MessageBoxOptions, Tray, app, dialog, nativeImage} from "electron";
 import {createInviteWindow, mainWindow} from "./discord/window";
-import * as path from "path";
+import path from "path";
 import {createSettingsWindow} from "./settings/main";
 import {getConfig, getConfigLocation, setConfig} from "./common/config";
 import {getDisplayVersion} from "./common/version";
-export let tray: any = null;
+export let tray: Tray;
 let trayIcon = "ac_plug_colored";
 void app.whenReady().then(async () => {
     // REVIEW - app will hang at startup if line above is awaited.
