+++ src/types/themeManifest.d.ts
@@ -1,0 +1,16 @@
+export interface ThemeManifest {
+    name: string;
+    author?: string;
+    description?: string;
+    version?: string;
+    invite?: string;
+    authorId?: string;
+    theme: string;
+    authorLink?: string;
+    donate?: string;
+    patreon?: string;
+    website?: string;
+    source?: string;
+    updateSrc?: string;
+    supportsArmCordTitlebar?: boolean;
+}
