+++ src/types/windowState.d.ts
@@ -1,0 +1,7 @@
+export interface WindowState {
+    width: number;
+    height: number;
+    x: number;
+    y: number;
+    isMaximized: boolean;
+}
